+++ src/main/java/com/ohgoodteam/ohgoodpay/shorts/dto/response/profile/ShortsProfileEditResponseDto.java
@@ -1,0 +1,6 @@
+package com.ohgoodteam.ohgoodpay.shorts.dto.response.profile;
+
+public record ShortsProfileEditResponseDto(
+    String success,
+    String message
+){}
